requirements.txt
yt-dlp
python-telegram-bot==13.15